+++ client/src/components/Navbar/index.jsx
@@ -1,5 +1,6 @@
 import { useAuth } from "../../hooks/auth";
 import { DefineRole } from "../../hooks/routes";
+import { useTitle } from "../../hooks/useTitle";
 import shipLogo from "../../assets/img/shipLogo.svg";
 
 const NavLinks = [
@@ -69,61 +70,65 @@ function Nav() {
     const { logout } = useAuth();
     const { original_role } = DefineRole();
 
+    const pageName = NavLinks.filter((nav) => {
+        return nav.link.includes(window.location.pathname);
+    });
+
+    useTitle(`KEL-BD | ${pageName[0]?.name}`);
+
     return (
         <div
-            className={`flex flex-wrap place-items-center ${
+            className={`flex w-full ${
                 (window.location.pathname === "/login" ||
                     window.location.pathname === "/") &&
                 "hidden"
             } `}
         >
-            <section className="relative mx-auto">
-                <nav className="flex justify-between rounded-b-md bg-green-700 text-white">
-                    <div className="flex w-full items-center px-5 py-2 xl:px-8">
-                        <a
-                            className="font-heading flex items-center justify-center text-2xl font-bold"
-                            href="/"
-                        >
-                            <img
-                                className="mr-1 h-10 w-10 text-white"
-                                src={shipLogo}
-                                alt="KEL-BD"
-                            />
-                            KEL-BD
-                        </a>
-                        <ul className="font-heading mx-auto flex space-x-12 px-4 font-semibold">
-                            {NavLinks.filter((nav) => {
-                                return nav.role.includes(original_role);
-                            }).map(({ id, link, name }) => {
-                                return (
-                                    <li key={id}>
-                                        <a
-                                            className={`rounded-md border-b-4 py-2 text-center transition duration-500 ease-in-out hover:rounded-md hover:border-b-4 hover:border-gray-200 hover:text-gray-200
+            <nav className="flex min-w-full justify-between rounded-b-md bg-green-700 text-white">
+                <div className="flex w-full items-center px-5 py-2 xl:px-8">
+                    <a
+                        className="font-heading flex items-center justify-center text-2xl font-bold"
+                        href="/"
+                    >
+                        <img
+                            className="mr-1 h-10 w-10 text-white"
+                            src={shipLogo}
+                            alt="KEL-BD"
+                        />
+                        KEL-BD
+                    </a>
+                    <ul className="font-heading mx-auto flex space-x-12 px-4 font-semibold">
+                        {NavLinks.filter((nav) => {
+                            return nav.role.includes(original_role);
+                        }).map(({ id, link, name }) => {
+                            return (
+                                <li key={id}>
+                                    <a
+                                        className={`rounded-md border-b-4 py-2 text-center transition duration-500 ease-in-out hover:rounded-md hover:border-b-4 hover:border-gray-200 hover:text-gray-200
                                             ${
                                                 window.location.pathname ===
                                                 link
                                                     ? "border-white"
                                                     : "border-green-700"
                                             }`}
-                                            href={link}
-                                        >
-                                            {name}
-                                        </a>
-                                    </li>
-                                );
-                            })}
-                        </ul>
-                        <div className="hidden items-center space-x-5 xl:flex ">
-                            <button
-                                onClick={logout}
-                                className="hover:border-red rounded-full border-2 border-red-600 bg-white px-6 py-2 font-sans font-semibold capitalize text-red-600 transition-all duration-500 ease-in-out hover:bg-red-600 hover:text-white"
-                            >
-                                {original_role}
-                            </button>
-                        </div>
+                                        href={link}
+                                    >
+                                        {name}
+                                    </a>
+                                </li>
+                            );
+                        })}
+                    </ul>
+                    <div className="hidden items-center space-x-2 xl:flex ">
+                        <button
+                            onClick={logout}
+                            className="hover:border-red rounded-full border-2 border-red-600 bg-white px-6 py-2 font-sans font-semibold capitalize text-red-600 transition-all duration-500 ease-in-out hover:bg-red-600 hover:text-white"
+                        >
+                            {original_role}
+                        </button>
                     </div>
-                </nav>
-            </section>
+                </div>
+            </nav>
         </div>
     );
 }
