+++ client/src/hooks/useTitle.jsx
@@ -1,0 +1,12 @@
+import { useEffect } from "react";
+
+export function useTitle(title) {
+    useEffect(() => {
+        const prevTitle = document.title;
+        document.title = title;
+
+        return () => {
+            document.title = prevTitle;
+        };
+    }, []);
+}
