+++ client/src/pages/PreDefinedAdmin/index.jsx
@@ -138,7 +138,6 @@ const App = () => {
                 LV_name: newStatus.LV_name, //handleAddFormChange로 받은 새 데이터
             }
         ).then((response) => {
-            console.log(response);
             generatedToast(response);
         });
 
@@ -186,7 +185,6 @@ const App = () => {
                 remark: editedStatus.remark,
             }
         ).then((response) => {
-            console.log(response);
             generatedToast(response);
         });
 
@@ -236,7 +234,6 @@ const App = () => {
                 status_id: StatusId,
             }
         ).then((response) => {
-            console.log(response);
             generatedToast(response);
         });
         newCurrentStatus.splice(index, 1);
